A few minor correction on handling paths to local build directories
git-svn-id: http://usg.lofar.org/svn/code/trunk@739 e3a606d5-b720-0410-b6ad-cb58f335abf8

+++ build/build.sh
@@ -61,13 +61,27 @@ cmake ../../external/hdf5
 
 ## [2] USG packages ----------------------------------------
 
-cd $basedir/dal
-rm -rf *
+cd $basedir
+if test -d dal ; then
+    cd dal
+    rm -rf *
+else 
+    mkdir dal
+    cd dal
+fi
 cmake ../../src/DAL
 make install
 
-cd $basedir/cr
-rm -rf *
+## CR-Tools
+
+cd $basedir
+if test -d cr ; then
+    cd cr
+    rm -rf *
+else 
+    mkdir cr
+    cd cr
+fi
 cmake ../../src/CR-Tools
 make && make install
 ctest
